toggle reaction if user already set it
