+++ client/src/features/reaction/ui/ReactionsContext.vue
@@ -24,6 +24,7 @@ const add = async (stickerId: string) => {
     if (!props.reactionPanelContext.message) return;
     await addReaction(props.reactionPanelContext.message, stickerId);
     props.reactionPanelContext.isVisible = false;
+    console.log(123);
 };
 </script>
